fix: seperate test and create a new channel
