+++ e2e/cypress/integration/channel/message_spec.js
@@ -8,7 +8,6 @@
 // ***************************************************************
 
 import users from '../../fixtures/users.json';
-import TIMEOUTS from '../../fixtures/timeouts.js';
 
 const sysadmin = users.sysadmin;
 
@@ -140,67 +139,5 @@ describe('Message', () => {
             // * Check that message has the css class needed for system message styling
             cy.get(divPostId).should('have.class', 'post--system');
         });
-    });
-
-    it('MM-16730 Reply to an older message', () => {
-        cy.getCurrentChannelId().then((channelId) => {
-            // # Get yesterdays date in UTC
-            const yesterdaysDate = Cypress.moment().subtract(1, 'days').valueOf();
-
-            // # Post a day old message
-            cy.postMessageAs({sender: sysadmin, message: 'Hello from yesterday', channelId, createAt: yesterdaysDate}).
-                its('id').
-                should('exist').
-                as('yesterdaysPost');
-        });
-
-        // # Add two subsequent posts
-        cy.postMessage('One');
-        cy.postMessage('Two');
-
-        cy.get('@yesterdaysPost').then((postId) => {
-            // # Scroll up to yesterdays message
-            cy.get(`#post_${postId}`).scrollIntoView();
-
-            // # Open RHS comment menu
-            cy.clickPostCommentIcon(postId);
-
-            // # Upload an attachment to the reply
-            cy.fileUpload('#fileUploadInput').wait(TIMEOUTS.TINY);
-
-            // # Reply with the attachment
-            cy.postMessageReplyInRHS('A reply to an older post with attachment');
-
-            // # Get the latest reply post
-            cy.getLastPostId().then((replyId) => {
-                // * Verify that the reply is in the channel view with matching text
-                cy.get(`#post_${replyId}`).within(() => {
-                    cy.queryByTestId('post-link').should('be.visible').and('have.text', 'Commented on sysadmin\'s message: Hello from yesterday');
-                    cy.get(`#postMessageText_${replyId}`).should('be.visible').and('have.text', 'A reply to an older post with attachment');
-                    cy.queryByTestId('image-name').should('be.visible').and('have.text', 'mattermost-icon.png');
-                });
-
-                // * Verify that the reply is in the RHS with matching text
-                cy.get(`#rhsPost_${replyId}`).within(() => {
-                    cy.queryByTestId('post-link').should('not.be.visible');
-                    cy.get(`#postMessageText_${replyId}`).should('be.visible').and('have.text', 'A reply to an older post with attachment');
-                    cy.queryByTestId('image-name').should('be.visible').and('have.text', 'mattermost-icon.png');
-                });
-
-                cy.get(`#CENTER_time_${postId}`).find('#localDateTime').invoke('attr', 'title').then((originalTimeStamp) => {
-                    // * Verify the first post timestamp equals the RHS timestamp
-                    cy.get(`#RHS_ROOT_time_${postId}`).find('#localDateTime').invoke('attr', 'title').should('be', originalTimeStamp);
-
-                    // * Verify the first post timestamp was not modified by the reply
-                    cy.get(`#CENTER_time_${replyId}`).find('#localDateTime').should('have.attr', 'title').and('not.equal', originalTimeStamp);
-                });
-            });
-        });
-
-        // # Close RHS
-        cy.closeRHS();
-
-        // # Verify RHS is open
-        cy.get('#rhsContainer').should('not.be.visible');
     });
 });
